+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - Support runtime projector names ([#32](https://github.com/commanded/commanded-ecto-projections/pull/32)).
+- Support `schema_prefix/2` function ([#33](https://github.com/commanded/commanded-ecto-projections/pull/33)).
 
 ---